Bugfix: Respect visibility of roles
Before this patch, the BBB role lists which are shown to the user contained all roles
which exist in Moodle. This is especially bad if a teacher is configuring a BBB
activity instance and sees roles which he should not be able to see.

This patch now makes sure that only the roles which are visible to the current user
are shown in the list.

Visibility of roles was added to Moodle core in Moodle 3.5 (See MDL-50666 for details).