Bugfix: "Moderator by default" setting was broken
Before this patch, the list which was shown to the admin in the "Moderator by default"
setting did not have the correct option values set. This was because in
in bigbluebuttonbn_settings_participants() the role array and the owner array were combined
with array_merge() which lead to renumbering of the array and destroyed the role IDs.

This patch now makes sure that the indexes are preserved and the setting is working
correctly.

This patch does not handle existing installations. If the setting should be fixed for
existing installations, an upgrade script will be needed in a subsequent patch.