CONTRIB-8053: List Table Pagination Bugfix  and CONTRIB-8052: Course List Improvement (#224) (#232)
* Bug fix for large lists

Bug fix for large lists. The link is written as "#". The page goes to the top when the link is clicked. But the confirm box remains at the bottom of the page and is not clicked.

* Add Shortname detail

* Add Shortname detail

* Recalc page count (Fixed)

* Update locallib.php

travis-ci error fixed: Inline control structures are not allowed

* travis-ci fix

Expected "if (...) {\n"; found "if(...){\n"

Co-authored-by: CAGLAR OFLAZOGLU <caglar@oflazoglu.org>