Let moderators receive the guest access link in the chat
This patch adds a new admin setting and activity instance setting which enables the 'Moderator receives guest link in chat' feature.
With this feature, moderators will receive the guest link in the chat of the conference room.
This way, they don't need to switch back to Moodle when they need to invite another participant.