CONTRIB-8596: Replace $bbsession array (#419)
Co-authored-by: Andrew Lyons <andrew.lyons@moodle.com>

* creating a new instance class with appropriate helpers and function to fetch the legacy $bbbsession and other data
* updating each endpoint one-by-one to use the instance
* updating all functions that the endpoint calls to use the instance
* rename/remove old functions as we go
* Adding cache where needed for information that is retrieved often from
the server.
* Update view.php to use new instance object
* Move server test to bigbluebutton class
* move bbb_view to use new instance
* move web services to use instanceid
* Update recording editables to use the instance
* Rewrite view.php to use a renderer properly
This change also includes changes to the meeting_info web service which
moves the generation of meeting information to a new class
`mod_bigbluebuttonbn\meeting`.
This change needs to be done at the same time so that the template for
the view_page can use the same variabels as when the meeting information
is fetched via the meeting_info and end_meeting web services, where it
is used to update the template.
The effect of this change means a simplification to the javascript for
joining a room.
* Update JS to not DOS server
The Room Updating code was polling the server every few seconds, and not
waiting for the previous poll request to finish. This leads to a backlog
of connections and a low grade DOS attack on the server.
* Fix presentation icon and display
* Only show recordings for own group
* Migrate import_view to use instance instead of bbbsession
* Use instance and try to remove as much as possible $bbbsession objects
* Refactor helper/instance into mod_helper (that can be then refactored
or included in other class)
* Cache the BBB server version (first step toward caching more
information). This is called often in the process of checking the server
so we need a way to cache this information.
* An administrator is also a moderator
* The now unused functions/method used to create $bbbsession have been removed
* Small bugs / code style fixing
* Removed function in broker.php that were not used anymore
* Check unit/behat tests
* The completionview event was not mentioned in the sort_order leading
to an exception on M3.11
* These were previously only used when generating the view page, which has
been converted to a renderer and uses CTA and notifications.
* Remove empty thirdpartylibs
* Fix various bugs in index page
* Deprecate and remove necurl
* Jump straight to group for grouped rooms
* Do not return an empty presentation
* Convert end session button link
A link arguably has better accessibility than a button as it harder to
press accidentally. Pressing [space] on a button submits it, but does
nothing on a link. For danger conditions we should therefore use a link,
because [space] often has other uses - e.g. page down.
* Fix opening + closing times
* The meeting class was incomplete and meeting_helper was redudant
* The responsability of contacting the BBB server is pushed down into
local\bigbluebutton class. This means we had also to refactor the
recording classes.
* Refactor meeting_helper and add new exceptions
* Remove meeting_helper and mobileview classes that are now unused
* Fix issues regarding recordings (and fetching recording)
* Fix various bugs in index page
* Fix issue with notifier

Co-authored-by: Andrew Nicols <andrew@nicols.co.uk>