CONTRIB-8660: Refactor notifier into proper notifications
The notifier class was creating messages which could not be proprerly
translated as they included concatenated strings to create complex
sentences.

It also generated HTML and stored this as custom data which was not
scalable. A large number of adhoc tasks were created from the frontend
rather than a single task being created which was responsible for the
remaining tasks.

The adhoc tasks were setting the recipient as customdata, rather than
running the task as that user.

This change breaks out the task into two separate tasks, and ensures
that notifications are sent whilst logged into the recipient, which
ensures that they use the correct language for that user in all
translated strings. Only one task is created from the frontend, and only
minimal data is stored. The user list is calculate during send to ensure
it is up-to-date.

Unfortunately it is not possible to migrate the legacy tasks to use the
new task API, but the behaviour is preserved so that existing tasks will
be able to run. No new tasks using the legacy notifier should be created
and this legacy class can be removed as per the deprecation policy.